+++ tests/e2e/setup/chain.go
@@ -41,9 +41,11 @@ type chain struct {
 	dataDir        string
 	ID             string
 	Validators     []*validator
+	TestAccounts   []*testAccount
 	GaiaValidators []*gaiaValidator
 }
 
+// newChain creates a chain with a random chain ID and makes a temporary directory to hold its data
 func newChain() (*chain, error) {
 	tmpDir, err := os.MkdirTemp("", "umee-e2e-testnet-")
 	if err != nil {
@@ -56,40 +58,44 @@ func newChain() (*chain, error) {
 	}, nil
 }
 
+// configDir is located at <chain.dataDir>/<chain.ID>
 func (c *chain) configDir() string {
 	return fmt.Sprintf("%s/%s", c.dataDir, c.ID)
 }
 
+// createAndInitValidators adds a number of validators to the chain and initializes their
+// keys, config.toml and genesis.json in separate config directories created for each validator.
+// Their monikers are all set to "umee" but their indexes are different.
 func (c *chain) createAndInitValidators(cdc codec.Codec, count int) error {
 	for i := 0; i < count; i++ {
-		node := c.createValidator(i)
-
-		// generate genesis files
-		if err := node.init(cdc); err != nil {
-			return err
+		val := validator{
+			chain:   c,
+			index:   i,
+			moniker: "umee",
 		}
 
-		c.Validators = append(c.Validators, node)
+		// create config directory and initializes config.toml and genesis.json
+		if err := val.init(cdc); err != nil {
+			return err
+		}
 
-		// create keys
-		if err := node.createKey(cdc, "val"); err != nil {
+		// generate a random key and save to keyring-backend-test in the validator's config directory
+		if err := val.createKey(cdc, "val"); err != nil {
 			return err
 		}
-		if err := node.createNodeKey(); err != nil {
+		// loads or generates a node key in the validator's config directory
+		// TODO (comment): which are we doing? loading or generating?
+		if err := val.createNodeKey(); err != nil {
 			return err
 		}
-		if err := node.createConsensusKey(); err != nil {
+		// loads or generates a consensus key in the validator's config directory
+		// TODO (comment): which are we doing? loading or generating?
+		if err := val.createConsensusKey(); err != nil {
 			return err
 		}
+
+		c.Validators = append(c.Validators, &val)
 	}
 
 	return nil
-}
-
-func (c *chain) createValidator(index int) *validator {
-	return &validator{
-		chain:   c,
-		index:   index,
-		moniker: "umee",
-	}
 }
